requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-database-url==1.0.3
django-taggit==3.0.0
djangorestframework==3.13.1
django-summernote==0.8.20.0
drf-jwt==1.19.2
gunicorn
mysqlclient==2.1.1
Pillow==9.2.0
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
djangorestframework-simplejwt==5.2.0
PyJWT==2.4.0
whitenoise
